src/requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
dnspython==1.16.0
idna==2.9
Jinja2==2.11.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
pkg-resources==0.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
soupsieve==2.0
stem==1.8.0
torrequest==0.1.0
urllib3==1.25.8
